feat(App.vue): add LanguageSelector component to handle language selection fix(App.vue): remove unused code related to locales and Combobox feat(LanguageSelector.vue): create LanguageSelector component for language selection
